+++ functions/src/index-new.ts
@@ -58,3 +58,4 @@ export * from './migration';
 export { guestPortalOps } from './guestPortal';
 
 
+