+++ functions/src/migration.ts
@@ -171,3 +171,4 @@ export const migrateToMultiGroup = functions.runWith({
 
 
 
+