+++ functions/src/shared/constants/index.ts
@@ -5,3 +5,4 @@ export const LOCATION = 'us-central1';
 export const QUEUE = 'calendar-sync';
 
 
+