+++ src/components/modals/BirthdayQuickActionsModal.tsx
@@ -164,3 +164,4 @@ export const BirthdayQuickActionsModal: React.FC<BirthdayQuickActionsModalProps>
   );
 };
 
+